Ignore TXT records
TXT records are buggy in the dns-sd library. There is an
error in the ByteArrayReader.prototype.getBytes method for
TXT records, where the caller expects a ByteArray object to
be returned (with its getReader()) method, but instead just a
Uint8Array is returned. Wrapping these returns with an empty
ByteArray leads to an incomplete parsing of the packet, as an
eof is not recorded. TXT records are not needed for merely
detecting a service and a port, so for now we are going to
just ignore TXT records.