Save MHTML to app storage via extension
Basic message passing is implemented here to allow the
extension to save an MHTML file to app storage space via
message passing.

The extension architecture relies on Gulp, rather than Grunt,
which isn't ideal and currently we aren't using browserify.
That shorcut should be remedied, but this is still a good
step.