Copy seed-element and include tests
We want to start incorporating Polymer into the project and
we want to include tests, because we are good people. This
commit adds the basics of a testing system for web
components and polymer that uses their seed-element example
repo as a starting point. Tests can be run using
wct test/polymer-ui/elements after running bower install.