Move app drawer to custom element cache-selector
It isn't yet completely clear if this is the right approach
to sharing data between both the drawer and the iron-pages
element. My hope is that this will allow us to modularize
content, but it isn't yet completely clear to me how we
would allow a dom-repeat element to go between both of them
successfully. We might need to add a function to
app-controller and bind both to a single variable or
something along those lines. I'd like to get as much business
logic as possible OUT of the Polymer components and just have
Polymer serve as a thin presentation layer on top.