Add clean:bundles task
vulcanize requires that we build our browserify bundles in
the app/ directory, as we are allowing vulcanize to build the
files from the app/ directory and write the output to the
dist/ directory. This then leaves the bundles in the app/
directory, which breaks our idea of having no compiled/built
code in the app/ directory. This adds a clean:bundles task
that removes these compiled files as part of the build
process.