Temporarily add incremental packet IDs
To facilitate wire-sharking, this commit adds incremental IDs
to DNS packets as they are sent. This makes it easy to
monitor when a packet is lost between two machines, but it is
not strictly compliant with the mDNS spec. This commit should
BE REVERTED once the issue is tracked down.