Machinery in place to run trials, not by hand
It looks like global variables from the contentscript do not
get added to the window object, which mean we can't just do
require('eval') or window.evaluation like I'd been doing in
the past. Will have to find a workaround. For now, however,
the machinery at least is in place.