Manual chunking implemented without tests
In the continuing saga to ensure that I understand the webrtc APIs, this
commit adds chunking but does not yet include tests. At the moment I am
still trying to make sure that this API behaves like I expect it to
behave. I did not realize that chunking needs to be performed manually,
eg.