Add API for getting list from service name
Previously the UI was doing some semi-complicated logic to resolve a
cache from a service name and then query that cache for a list. Adding
webrtc means we have to accommodate the new PeerAccessor interface.
Rather than keeping this in the Polymer code, this commit adds a
function to get a list from a service name using a single API.

This continues to heighten my concern that the app-controller module is
becoming a bit of a god object. I'd like to consider moving some of
these API and function-composing calls into a separate module, and keep
app-controller only for starting and controlling app/service state. For
now, however, I'm comfortable leaving it this way to keep things
incremental.