Add tests for chrome-apis/file-system
Starting to add tests for the Chrome family of APIs to make sure we are
rejecting Promises as expected in response to errors.

It isn't clear if these APIs need to catch a throw or if they just need
to check the last chrome.runtime.lastError property. Might have to
rework some of these. Just getting a schema in place atm.