Fix typo in chromep, works with one computer
Testing on a single machine and this seems to work as I would expect
after more fully moving to chromep. This has the benefit of removing
some of the code under my control, but I haven't tested it with error
conditions yet to the point to verify if it works as expected.

For instance, looking at the code it isn't clear to me if it is
differentiating between errors generated by setting
chrome.runtime.lastError and those API calls that throw a runtime error.
As I was switching my own to the chrome-promise like solution, before
finding chrome-promise, I found some API calls I was using that could
perhaps be using runtime errors to handle exceptions. Thus for now,
chrome-promise is a way to reduce the test surface of our code but is
not yet a completely verified and known quantity.