Start moving chunking-channel to use protocol
Only the success case is handled at the moment. There is also a change
here to ensure that we require('buffer/'), rather than
require('buffer'). The require('buffer') method seemed to allow the
default Node Buffer module, which led to a weird API disparity where
readUInt32() required an offset in the Node implementation that was
resolved by Browserify, but not by the Buffer module that was being used
during tests. Very puzzling, but seem to have fixed it by the trailing
slash.