Digest coalescence strategy works locally
I'd missed a place to implement reponding to message in the webrtc
responder correctly. That is fixed, as are a couple other issues, and
we're now returning the correct information for our own caches. That is
trouble, though, as we don't want to annotate ourselves via the
coalescence manager, at least not as currently implemented.