Don't add br0 interface to network interfaces
This was one of the most puzzling bugs I've ever seen. I was trying to
talk to another machine on the network but was somehow talking to
myself. I think it turned out to be because the chromebooks expose an
external-looking ipv4 interface called 'br0'. This appeared to be a
localhost equivalent, like 127.0.0.1. Both chromebooks were advertising
this via mDNS, and then when connecting to it they were actually talking
to their own br0 interface. So you see the other machine as expected but
end up dialing yourself. So strange. I just remove br0 now to try and
prevent this.