Add paging, polymer tests breaking b/c databinding
It looks like there is an open bug on Polymer 2.0 and wct 6 that data
binding is broken:

https://github.com/PolymerElements/test-fixture/issues/47

This is a huge problem, and I don't understand why it remains open after
being opened in March. However, the functionality is important enough
that, for now, I am going to begrudgingly allow it to break.