Change peer-interface to client, drop http
Peer-accessor is really a client that talks to the server module. I
don't know why I named it so strangely originally. The http-impl
interface was also a pain in the butt, as it wasn't as efficient as the
webrtc implementation. As we moved to Buffers rather than JSON, this
also made the http-impl less useful for debugging, as now it was just
showing buffer blobs. So we dropped it.